+++ api/controllers/ContactController.js
@@ -12,13 +12,4 @@ module.exports = {
       res.serverError(e);
     }
   },
-  show: async (req, res) => {
-    try {
-      let data = await Post.findByIdHasJoin(req.params.id);
-      const social = SocialService.forPost({posts: [data]});
-      res.view('blog/show', {data, social});
-    } catch (e) {
-      res.serverError(e);
-    }
-  },
 }