+++ api/controllers/api/admin/AllpayController.js
@@ -1,4 +1,3 @@
-import allPayPaymentTypeJson from '../../../../config/allpayPaymentType.json';
 
 module.exports = {
 
@@ -116,67 +115,6 @@ module.exports = {
 
       const message = 'Delete success.';
       res.ok({ message, data: { allpay, order } });
-    } catch (e) {
-      res.serverError(e);
-    }
-  },
-
-  paid: async (req, res) => {
-    try {
-      const data = req.body;
-      sails.log.info(data);
-      const allpay = await AllpayService.paid(data);
-
-      //  create and send message
-      let messageConfig = {};
-      messageConfig.serialNumber = allpay.TradeNo;
-      if (allpay.RecipeOrderId) {
-        const recipeOrder = await RecipeOrder.findByIdHasJoin(allpay.RecipeOrderId);
-        recipeOrder.productionStatus = 'PAID';
-        await recipeOrder.save();
-        messageConfig.email = recipeOrder.email;
-        messageConfig.username = recipeOrder.User.displayName;
-      }
-      messageConfig = await MessageService.paymentConfirm(messageConfig);
-      const message = await Message.create(messageConfig);
-      await MessageService.sendMail(message);
-
-      res.send('1|OK');
-    } catch (e) {
-      res.serverError(e);
-    }
-  },
-
-  paymentinfo: async(req, res) => {
-    try {
-      const data = req.body;
-      sails.log.info(data);
-      const allpay = await AllpayService.paymentinfo(data);
-
-      //  create and send message
-      let messageConfig = {};
-      messageConfig.serialNumber = allpay.TradeNo;
-      messageConfig.paymentTotalAmount = allpay.ShouldTradeAmt;
-      messageConfig.bankName = allPayPaymentTypeJson[allpay.PaymentType] || allpay.PaymentType;
-      messageConfig.bankId = allpay.BankCode;
-      messageConfig.accountId = allpay.vAccount;
-      messageConfig.expireDate = allpay.ExpireDate;
-      if (allpay.RecipeOrderId) {
-        const recipeOrder = await RecipeOrder.findByIdHasJoin(allpay.RecipeOrderId);
-        messageConfig.productName = recipeOrder.Recipe.perfumeName + ' 100 ml';
-        messageConfig.email = recipeOrder.email;
-        messageConfig.username = recipeOrder.User.displayName;
-        messageConfig.shipmentUsername = recipeOrder.recipient;
-        messageConfig.shipmentAddress = recipeOrder.address;
-        messageConfig.note = recipeOrder.note;
-        messageConfig.phone = recipeOrder.phone;
-        messageConfig.invoiceNo = recipeOrder.invoiceNo;
-      }
-      messageConfig = await MessageService.orderConfirm(messageConfig);
-      const message = await Message.create(messageConfig);
-      await MessageService.sendMail(message);
-
-      res.send('1|OK');
     } catch (e) {
       res.serverError(e);
     }